+++ src/main/java/com/everypet/member/mapper/MsMemberInfoMapper.java
@@ -1,6 +1,6 @@
 package com.everypet.member.mapper;
 
-import com.everypet.member.model.dto.MemberInfoDTO;
+import com.everypet.member.model.dto.member.MemberInfoDTO;
 import com.everypet.member.model.vo.Member;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
@@ -10,7 +10,5 @@ public interface MsMemberInfoMapper {
     MsMemberInfoMapper INSTANCE = Mappers.getMapper(MsMemberInfoMapper.class);
 
     MemberInfoDTO toDto(Member vo);
-
-    Member toVo(MemberInfoDTO dto);
 
 }
