+++ src/main/java/com/everypet/member/mapper/MsPasswordRecovery.java
@@ -1,6 +1,6 @@
 package com.everypet.member.mapper;
 
-import com.everypet.member.model.dto.PasswordRecoveryDTO;
+import com.everypet.member.model.dto.member.PasswordRecoveryDTO;
 import com.everypet.member.model.vo.Member;
 import com.everypet.member.model.vo.PasswordRecovery;
 import org.mapstruct.Mapper;
