+++ src/main/java/com/everypet/member/mapper/MsSignupMapper.java
@@ -1,6 +1,6 @@
 package com.everypet.member.mapper;
 
-import com.everypet.member.model.dto.SignupDTO;
+import com.everypet.member.model.dto.member.SignupDTO;
 import com.everypet.member.model.vo.Member;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
@@ -9,8 +9,6 @@ import org.mapstruct.factory.Mappers;
 public interface MsSignupMapper {
 
     MsSignupMapper INSTANCE = Mappers.getMapper(MsSignupMapper.class);
-
-    SignupDTO toDto(Member vo);
 
     Member toVo(SignupDTO dto);
