+++ src/main/java/com/everypet/member/model/dto/address/AddressRegisterDTO.java
@@ -1,4 +1,4 @@
-package com.everypet.member.model.dto;
+package com.everypet.member.model.dto.address;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AddressDTO {
+public class AddressRegisterDTO {
 
     @ApiModelProperty(example = "서울시 강남구 롯데캐슬 아파트", notes = "주소")
     private String address; // 주소
