+++ src/main/java/com/everypet/member/model/dto/address/SignupAddressDTO.java
@@ -1,6 +1,5 @@
-package com.everypet.member.model.dto;
+package com.everypet.member.model.dto.address;
 
-import com.everypet.member.model.vo.Address;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,17 +21,5 @@ public class SignupAddressDTO {
     @ApiModelProperty(example = "105-1505", notes = "상세 주소")
     @NotBlank(message = "상세 주소는 필수 입력 값입니다.")
     private String detailAddress;
-
-    public Address toEntity(SignupDTO member) {
-        return Address.builder()
-                .memberId(member.getMemberId())
-                .address(address)
-                .detailAddress(detailAddress)
-                .receiver(member.getMemberId())
-                .phone(member.getPhone())
-                .request("문 앞")
-                .defaultYn('Y')
-                .build();
-    }
 
 }
